feat: Leave backticks unencoded in query strings where possible
Addresses #1397.

When the user provides (part of) their query as a string, we leave any
unencoded and encoded backticks in the string exactly as is, trusting
that the user has encoded properly.

When the user calls req.query(Object), and one of the keys/values in
Object has an unencoded backtick, we now don't encode that either. This
is consistent with the WHATWG specification and is as close as we can
get to being conceptually consistent with the `req.query(String)` form.

A fundamental limit of the `req.query(Object)` format is that the user
can't indicate "a backtick meant as a delimiter" vs. "a backtick meant
as data”; this limitation becomes apparent when you consider what
happens if the user tries to provide an already encoded backtick in
their query object, in an attempt to distinguish these two cases,
namely: the % sign gets encoded again to %25, like any other object
character needing encoding. And there's no way around that/that’s how
it should be. (Consider that all other potential delimiter characters
that can appear unencoded have the same behavior. There’s just no way,
in the object form, to know what the user wants.)

The new test in `test/client/serialize.js` is to recognize that, even
though we’re now following WHATWG rules on the server side, the
serialize function tested in that file is a serializer for
`application/x-www-form-urlencoded`, and its spec is unambiguous that
backticks get percent-encoded.